+++ Tests/ASCCommandTests/Commands/SubscriptionGroups/SubscriptionGroupsListTests.swift
@@ -22,7 +22,7 @@ struct SubscriptionGroupsListTests {
             {
               "affordances" : {
                 "createLocalization" : "asc subscription-group-localizations create --group-id grp-1 --locale en-US --name <name>",
-                "createSubscription" : "asc subscriptions create --group-id grp-1 --name <name> --product-id <id> --period ONE_MONTH",
+                "createSubscription" : "asc subscriptions create --group-id grp-1 --name <name> --period ONE_MONTH --product-id <id>",
                 "delete" : "asc subscription-groups delete --group-id grp-1",
                 "listLocalizations" : "asc subscription-group-localizations list --group-id grp-1",
                 "listSubscriptions" : "asc subscriptions list --group-id grp-1",